Upload of all project code files and data sets
The Java code folder holds all of the .java and .class files.  These Java files contain the necessary code for the logistic regression model algorithm and the Hadoop Map-Reduce framework.  The R code folder contains the data preprocessing code used on the datasets.  The datasets folder contains the training and testing datasets used on the logistic regression model.